Separate public cert methods from token decode methods
