+++ firebase_ruby_auth.gemspec
@@ -1,14 +1,17 @@
 # frozen_string_literal: true
 
+$LOAD_PATH.unshift File.expand_path('lib', __dir__)
+
 Gem::Specification.new do |s|
   s.name        = 'firebase_ruby_auth'
-  s.version     = '0.0.4'
+  s.version     = '0.0.5'
   s.date        = '2019-05-30'
   s.summary     = 'Authenticate Firebase User Tokens in Ruby'
   s.description = 'Firebase Ruby Auth'
   s.authors     = ['Emily Ring']
   s.email       = 'railsclt@gmail.com'
-  s.files       = ['lib/firebase_ruby_auth.rb']
+  s.files       = `git ls-files -- lib/*`.split("\n")
+  s.test_files  = `git ls-files -- {spec}/*`.split("\n")
   s.homepage    = 'https://github.com/railscltgroup/firebase_auth'
   s.license     = 'MIT'
