+++ lib/firebase_ruby_auth.rb
@@ -3,27 +3,20 @@
 # https://github.com/jwt/ruby-jwt
 require 'jwt'
 
-# https://github.com/typhoeus/typhoeus
-require 'typhoeus'
+require 'google_public_cert'
 
 # Interacts with data from Firebase
 class FirebaseRubyAuth
-  # This url is from the Google instructions:
-  # https://firebase.google.com/docs/auth/admin/verify-id-tokens
-  # This is were you can find the Google Public Key
-  CERT_URL = 'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com'
-
   def initialize(firebase_project_id)
     @firebase_project_id = firebase_project_id
-    fetch_google_public_key
+    @public_cert = GooglePublicCert.new
   end
 
   # token would be a user's ID token
   # https://firebase.google.com/docs/auth/admin/verify-id-tokens
   # This will either return a hash with user data, or an empty hash
   def decode_token(token)
-    fetch_google_public_key if @expires < Time.now
-    return {} if @keys.empty?
+    return {} if @public_cert.keys.empty?
 
     token_values = begin
                      JWT.decode(token, nil, true, options).first
@@ -33,25 +26,6 @@ class FirebaseRubyAuth
                      {}
                    end
     valid?(token_values) ? token_values : {}
-  end
-
-  private def fetch_google_public_key
-    # This requires the Typhoeus gem: https://github.com/typhoeus/typhoeus
-    request = Typhoeus.get(CERT_URL)
-    generate_keys(request)
-    generate_key_expiry(request)
-  end
-
-  private def generate_keys(request)
-    @keys = {
-      keys: (JSON.parse request.body).map do |key, value|
-        JWT::JWK.new(OpenSSL::X509::Certificate.new(value).public_key)
-                .export
-                .merge(kid: key)
-      end
-    }
-  rescue JSON::ParserError
-    @keys = {}
   end
 
   private def options
@@ -66,17 +40,8 @@ class FirebaseRubyAuth
       iss: "https://securetoken.google.com/#{@firebase_project_id}",
       verify_iss: true,
 
-      jwks: @keys
+      jwks: @public_cert.keys
     }
-  end
-
-  private def generate_key_expiry(request)
-    @expires = Time.new(request
-      .headers['cache-control']
-      .split(', ')
-      .select { |s| s.include?('max-age') }[0]
-      .split('max-age=')[1]
-      .to_i)
   end
 
   private def valid?(token_values)
